refactor(common): consolidate SrcRect + parse_src_rect into ooxml-common
The DrawingML `<a:srcRect>` source-image crop (ECMA-376 §20.1.8.55) was
implemented three times — pptx, docx, and xlsx each carried an identical
`SrcRect { l, t, r, b }` struct and a near-identical `parse_src_rect`
(edge / 100000, default 0, all-zero ⇒ None). Move both into
`ooxml_common::blip`, next to the sibling blip helpers (`blip_embed_rid`,
`svg_blip_rid`, `mime_from_ext`) the three parsers already share, and have
each parser import them.

- The shared `SrcRect` derives Serialize+Deserialize+Default+PartialEq+Copy so
  every consumer's needs are met. pptx's per-field `skip_serializing_if`
  (which omitted zero edges) is dropped — the `Option<SrcRect>` wrapper already
  omits the no-crop case, and all four edges now serialize uniformly (the TS
  side already tolerates this).
- The shared `parse_src_rect` takes the `<*:blipFill>` node; docx's call sites,
  which had the `<a:blip>`, pass `blip.parent()`.
- pptx keeps its local `is_zero_f64` (still used by `FillRect`/`parse_fill_rect`).

No behavior change: 67 xlsx + 100 pptx + 154 docx parser tests pass,
clippy --workspace clean.
