+++ packages/docx/parser/src/types.rs
@@ -1064,21 +1064,9 @@ pub struct ShapeText {
     pub image_height_pt: f64,
 }
 
-/// ECMA-376 §20.1.8.55 `<a:srcRect>` — the source rectangle that selects which
-/// region of the blip's bitmap is drawn (the rest is cropped away). The four
-/// attributes are ST_Percentage insets expressed here as FRACTIONS 0..1 of the
-/// source bitmap, measured inward from each edge: `l`/`t` from the left/top,
-/// `r`/`b` from the right/bottom. The visible source rectangle is therefore
-/// `[l, t, 1−r, 1−b]` of the bitmap. The parser converts the raw 1000ths-of-a-
-/// percent attribute values to fractions (÷100000) so the renderer needs no
-/// unit knowledge. An absent `<a:srcRect>` (or one that is all-zero) ⇒ `None`.
-#[derive(Serialize, Debug, Clone, PartialEq)]
-pub struct SrcRect {
-    pub l: f64,
-    pub t: f64,
-    pub r: f64,
-    pub b: f64,
-}
+/// ECMA-376 §20.1.8.55 `<a:srcRect>` source-image crop, shared across the docx,
+/// pptx and xlsx parsers (see `ooxml_common::blip`).
+pub use ooxml_common::blip::SrcRect;
 
 #[derive(Serialize, Debug, Clone)]
 #[serde(rename_all = "camelCase")]
